docker/silver_dbt/requirements.txt
dbt-duckdb==1.9.4
boto3==1.40.18
pyyaml==6.0.2
